requirements.txt
waitress
flask
requests
python-magic